backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.0
pydicom==2.4.4
torch==2.2.2
torchvision==0.17.2
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
